Move test classes into backend.config and backup.database packages
